refactor : productList에 관련된 함수 정리
- productList에 관련된 함수를 productList.js 파일에 담음
