+++ public/javascripts/main.js
@@ -1,4 +1,5 @@
 import { createNewElement, $ } from './utils.js';
+import { createProductListTemplate } from './productList.js';
 
 const productItemList = [
   {
@@ -37,37 +38,6 @@ const productItemList = [
     tagList: ['테마'],
   },
 ];
-
-const createTagListTemplate = tagList => {
-  const tagReducer = (acc, tag) =>
-    acc + `<div class="product-tag">${tag}</div>`;
-  return tagList.reduce(tagReducer);
-};
-
-const createProductListTemplate = productItemList => {
-  const productListContent = productItemList.reduce(
-    (acc, productItem) =>
-      acc +
-      `<li class="product-item">
-        <img
-          class="product-item__image"
-          src="${productItem.image}"
-        />
-        <div class="product-item__title">${productItem.title}</div>
-        <div class="product-item__description">
-          ${productItem.description}
-        </div>
-        <div class="product-tag-list">
-          <div class="product-tag">
-          ${createTagListTemplate(productItem.tagList)}
-          </div>
-        </div>
-      </li>`,
-    ''
-  );
-
-  return `<ul class="product-list__row">${productListContent}</ul>`;
-};
 
 const init = () => {
   const productList = $('.product-list');
